==> Scripts/07_GPC_Bulk_RNAseq_Comparison.Rmd <==
---
title: "Comparison of ESC- and iPSC-derived hGPCs to primary and adult hGPCs"
author: "John Mariani"
date: "10/11/23"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

```


```{r, message = F}
library(ggplot2)
library(tximport)
library(dplyr)
library(plyr)
library(tidyr)

```

### Load prior data
```{r}

ensemblGeneListH <- read.csv("data_for_import/ensemblGeneList.csv")
de_intersect <- read.delim("output/de_Adult_vs_Fetal_Intersect.txt")


```

### Comparison to iPSC line C27
```{r}

#Read in RSEM gene output
sampleTableGPC <- read.csv("data_for_import/sampleTableGPC.csv")
nrow(sampleTableGPC)

temp = list.files(path = "./data_for_import/genes", pattern="genes.results")
length(temp)

names(temp) <- substr(temp, 1, nchar(temp)-19)
temp <- temp[names(temp) %in% sampleTableGPC$sample]



txi.rsem.gpc <- tximport(paste0("./data_for_import/genes/",temp), type = "rsem")

for(i in 1:3){
  colnames(txi.rsem.gpc[[i]]) <- names(temp)
}


sampleTableGPC <- sampleTableGPC[match(names(temp), sampleTableGPC$sample),]


```


### GPC Heatmap

```{r, message=FALSE, warning=FALSE}

markerHMgenes <- c("MKI67", "CDK1", "TOP2A", "PDGFRA", "PTPRZ1", "LMNB1", "MYC", "TEAD2", "NFIB", "HMGA2", "HDAC2", "EZH2", "BCL11A", "MBP", "MAG", "MOG", "TF", "AHR", "IL1A","STAT3", "E2F6", "MAX", "ZNF274", "IKZF3")



markerTPM <- merge(txi.rsem.gpc$abundance, ensemblGeneListH, by.x = 0, by.y = "ensembl_gene_id")
markerTPM <- markerTPM[markerTPM$external_gene_name %in% markerHMgenes,]


markerTPM <- markerTPM[,1:31] %>%
  pivot_longer(-c(Row.names, external_gene_name), names_to = "Sample", values_to = "TPM")

markerTPM$group <- mapvalues(markerTPM$Sample,sampleTableGPC$sample, as.character(sampleTableGPC$label))

markerTPM$TPM <- log2(markerTPM$TPM + .1)

markerTPM$group <- factor(markerTPM$group, levels =c("ESC CD140a", "iPSC CD140a", "Fetal CD140a", "Fetal A2B5", "Adult A2B5"))
markerTPM$external_gene_name <- factor(markerTPM$external_gene_name, levels = rev(c(markerHMgenes)))


###

gpcHMgg <- ggplot(markerTPM, aes(Sample, external_gene_name)) + geom_tile(aes(fill = TPM)) + 
  theme(panel.spacing.y = unit(.5, "lines"), 
        panel.spacing.x = unit(.25,"lines"), 
        axis.title.x = element_blank(), 
        axis.title.y = element_blank(), 
        axis.text.x = element_blank(), 
        axis.ticks.y = element_blank(), 
        axis.text.y = element_text(colour = ifelse(levels(markerTPM$external_gene_name) %in% de_intersect[de_intersect$log2FoldChange > 0,]$external_gene_name, "#FF2020", "#2E30FF")))  + 
  scale_fill_gradientn(colours = c("#009900","#fffcbd","#ff2020"), guide = guide_colourbar(direction = "horizontal", title = "Log2(TPM + .1)", title.position = "top")) + scale_x_discrete(expand = c(0, 0))  + theme(axis.ticks.x = element_blank(), legend.position = "bottom", legend.direction = "horizontal") + facet_grid(cols = vars(group), scales = "free", space  = "free")


gpcHMgg

#ggsave("figures/ExFig4/gpcHMgg.pdf", units = "in", dpi = 300, width = 16, height = 8, device = NULL)



```

## Source Data
```{r}

sdS4A <- markerTPM
sdS4A <- sdS4A[,c(2:4)]

sdS4A <- pivot_wider(data = sdS4A,  names_from = Sample, id_cols = external_gene_name, values_from = TPM)
names(sdS4A)[2:4] <- paste0("Adult_", names(sdS4A)[2:4])
names(sdS4A)[1] <- "Gene"

sdS4A <- as.data.frame(sdS4A)

#write.xlsx(sdS4A, file = "Source Data/Source_Data_FigS4.xlsx", sheetName = "FigS4A", row.names = F)


```

```{r}

sessionInfo()

```
